as suggestion

+++ src/logger.ts
@@ -214,9 +214,14 @@ export function timeLoggedMethod<This, Args extends any[], Return>(
 //古いログファイルの削除
 
 const logger = getLogger("Logger");
-const deleteFiles = fs.readdirSync(path.join(__dirname, "../logs/"), { withFileTypes: true }).filter(d => d.isFile() && d.name.endsWith(".log")).map(d => d.name).sort().slice(0, -maxLogFiles);
+const deleteFiles = fs.readdirSync(path.join(__dirname, "../logs/"), { withFileTypes: true })
+  .filter(d => d.isFile() && d.name.endsWith(".log"))
+  .map(d => d.name)
+  .sort()
+  .slice(0, -maxLogFiles);
 
-logger.debug("Deleted " + deleteFiles.length + " log files.");
+if (deleteFiles.length > 0){
+  logger.debug("Deleted " + deleteFiles.length + " log files.");
 
-deleteFiles.forEach(name => fs.unlinkSync(path.join(__dirname, "../logs", name)));
-
+  deleteFiles.forEach(name => fs.unlinkSync(path.join(__dirname, "../logs", name)));
+}
